refactor(frontend): decompose Scanner.tsx into Scanner/ directory
